Move conversion utility to Schema
