+++ pinot-spi/src/test/java/org/apache/pinot/spi/data/DateTimeFieldSpecUtilsTest.java
@@ -16,17 +16,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.spi.utils;
+package org.apache.pinot.spi.data;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.concurrent.TimeUnit;
-import org.apache.pinot.spi.data.DateTimeFieldSpec;
 import org.apache.pinot.spi.data.FieldSpec.DataType;
-import org.apache.pinot.spi.data.TimeFieldSpec;
-import org.apache.pinot.spi.data.TimeGranularitySpec;
 import org.testng.Assert;
-import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 
@@ -47,13 +41,13 @@ public class DateTimeFieldSpecUtilsTest {
     timeFieldSpec = new TimeFieldSpec(new TimeGranularitySpec(DataType.LONG, TimeUnit.MILLISECONDS, "incoming"));
     expectedDateTimeFieldSpec =
         new DateTimeFieldSpec("incoming", DataType.LONG, "1:MILLISECONDS:EPOCH", "1:MILLISECONDS");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // incoming epoch hours
     timeFieldSpec = new TimeFieldSpec(new TimeGranularitySpec(DataType.INT, TimeUnit.HOURS, "incoming"));
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("incoming", DataType.INT, "1:HOURS:EPOCH", "1:HOURS");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // Simple date format
@@ -61,7 +55,7 @@ public class DateTimeFieldSpecUtilsTest {
         new TimeGranularitySpec(DataType.INT, TimeUnit.DAYS, "SIMPLE_DATE_FORMAT:yyyyMMdd", "incoming"));
     expectedDateTimeFieldSpec =
         new DateTimeFieldSpec("incoming", DataType.INT, "1:DAYS:SIMPLE_DATE_FORMAT:yyyyMMdd", "1:DAYS");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // simple date format STRING
@@ -69,13 +63,13 @@ public class DateTimeFieldSpecUtilsTest {
         new TimeGranularitySpec(DataType.STRING, TimeUnit.DAYS, "SIMPLE_DATE_FORMAT:yyyy-MM-dd hh-mm-ss", "incoming"));
     expectedDateTimeFieldSpec =
         new DateTimeFieldSpec("incoming", DataType.STRING, "1:DAYS:SIMPLE_DATE_FORMAT:yyyy-MM-dd hh-mm-ss", "1:DAYS");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // time unit size
     timeFieldSpec = new TimeFieldSpec(new TimeGranularitySpec(DataType.LONG, 5, TimeUnit.MINUTES, "incoming"));
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("incoming", DataType.LONG, "5:MINUTES:EPOCH", "5:MINUTES");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // transform function
@@ -83,7 +77,7 @@ public class DateTimeFieldSpecUtilsTest {
     timeFieldSpec.setTransformFunction("toEpochHours(timestamp)");
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("incoming", DataType.INT, "1:HOURS:EPOCH", "1:HOURS");
     expectedDateTimeFieldSpec.setTransformFunction("toEpochHours(timestamp)");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     /* 2] incoming + outgoing */
@@ -92,7 +86,7 @@ public class DateTimeFieldSpecUtilsTest {
     timeFieldSpec = new TimeFieldSpec(new TimeGranularitySpec(DataType.LONG, TimeUnit.HOURS, "time"),
         new TimeGranularitySpec(DataType.LONG, TimeUnit.HOURS, "time"));
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("time", DataType.LONG, "1:HOURS:EPOCH", "1:HOURS");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // same incoming and outgoing - simple date format
@@ -101,7 +95,7 @@ public class DateTimeFieldSpecUtilsTest {
             new TimeGranularitySpec(DataType.LONG, TimeUnit.DAYS, "SIMPLE_DATE_FORMAT:yyyyMMdd", "time"));
     expectedDateTimeFieldSpec =
         new DateTimeFieldSpec("time", DataType.LONG, "1:DAYS:SIMPLE_DATE_FORMAT:yyyyMMdd", "1:DAYS");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // millis to hours
@@ -109,7 +103,7 @@ public class DateTimeFieldSpecUtilsTest {
         new TimeGranularitySpec(DataType.LONG, TimeUnit.HOURS, "outgoing"));
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("outgoing", DataType.LONG, "1:HOURS:EPOCH", "1:HOURS");
     expectedDateTimeFieldSpec.setTransformFunction("toEpochHours(incoming)");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // millis to bucketed minutes
@@ -117,7 +111,7 @@ public class DateTimeFieldSpecUtilsTest {
         new TimeGranularitySpec(DataType.LONG, 10, TimeUnit.MINUTES, "outgoing"));
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("outgoing", DataType.LONG, "10:MINUTES:EPOCH", "10:MINUTES");
     expectedDateTimeFieldSpec.setTransformFunction("toEpochMinutesBucket(incoming, 10)");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // days to millis
@@ -125,7 +119,7 @@ public class DateTimeFieldSpecUtilsTest {
         new TimeGranularitySpec(DataType.LONG, TimeUnit.MILLISECONDS, "outgoing"));
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("outgoing", DataType.LONG, "1:MILLISECONDS:EPOCH", "1:MILLISECONDS");
     expectedDateTimeFieldSpec.setTransformFunction("fromEpochDays(incoming)");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // bucketed minutes to millis
@@ -133,7 +127,7 @@ public class DateTimeFieldSpecUtilsTest {
         new TimeGranularitySpec(DataType.LONG, TimeUnit.MILLISECONDS, "outgoing"));
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("outgoing", DataType.LONG, "1:MILLISECONDS:EPOCH", "1:MILLISECONDS");
     expectedDateTimeFieldSpec.setTransformFunction("fromEpochMinutesBucket(incoming, 5)");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // hours to days
@@ -141,7 +135,7 @@ public class DateTimeFieldSpecUtilsTest {
         new TimeGranularitySpec(DataType.INT, TimeUnit.DAYS, "outgoing"));
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("outgoing", DataType.INT, "1:DAYS:EPOCH", "1:DAYS");
     expectedDateTimeFieldSpec.setTransformFunction("toEpochDays(fromEpochHours(incoming))");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // minutes to hours
@@ -149,7 +143,7 @@ public class DateTimeFieldSpecUtilsTest {
         new TimeGranularitySpec(DataType.LONG, TimeUnit.HOURS, "outgoing"));
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("outgoing", DataType.LONG, "1:HOURS:EPOCH", "1:HOURS");
     expectedDateTimeFieldSpec.setTransformFunction("toEpochHours(fromEpochMinutes(incoming))");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // bucketed minutes to days
@@ -157,7 +151,7 @@ public class DateTimeFieldSpecUtilsTest {
         new TimeGranularitySpec(DataType.LONG, TimeUnit.DAYS, "outgoing"));
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("outgoing", DataType.LONG, "1:DAYS:EPOCH", "1:DAYS");
     expectedDateTimeFieldSpec.setTransformFunction("toEpochDays(fromEpochMinutesBucket(incoming, 10))");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // seconds to bucketed minutes
@@ -165,14 +159,14 @@ public class DateTimeFieldSpecUtilsTest {
         new TimeGranularitySpec(DataType.LONG, 5, TimeUnit.MINUTES, "outgoing"));
     expectedDateTimeFieldSpec = new DateTimeFieldSpec("outgoing", DataType.LONG, "5:MINUTES:EPOCH", "5:MINUTES");
     expectedDateTimeFieldSpec.setTransformFunction("toEpochMinutesBucket(fromEpochSeconds(incoming), 5)");
-    actualDateTimeFieldSpec = DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+    actualDateTimeFieldSpec = Schema.convertToDateTimeFieldSpec(timeFieldSpec);
     Assert.assertEquals(actualDateTimeFieldSpec, expectedDateTimeFieldSpec);
 
     // simple date format to millis
     timeFieldSpec = new TimeFieldSpec(new TimeGranularitySpec(DataType.LONG, TimeUnit.DAYS, "SIMPLE_DATE_FORMAT:yyyyMMdd", "incoming"),
         new TimeGranularitySpec(DataType.LONG, TimeUnit.MILLISECONDS, "outgoing"));
     try {
-      DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+      Schema.convertToDateTimeFieldSpec(timeFieldSpec);
       Assert.fail();
     } catch (Exception e) {
       // expected
@@ -182,7 +176,7 @@ public class DateTimeFieldSpecUtilsTest {
     timeFieldSpec = new TimeFieldSpec(new TimeGranularitySpec(DataType.LONG, TimeUnit.HOURS, "incoming"),
         new TimeGranularitySpec(DataType.INT, TimeUnit.HOURS, "SIMPLE_DATE_FORMAT:yyyyMMddhh", "outgoing"));
     try {
-      DateTimeFieldSpecUtils.convertToDateTimeFieldSpec(timeFieldSpec);
+      Schema.convertToDateTimeFieldSpec(timeFieldSpec);
       Assert.fail();
     } catch (Exception e) {
       // expected
